+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade to `stripes` `4.0`, `react-intl` `4.5`. Refs STRIPES-672.
 * Fixed label in move hold modal to show item title.  Fixes UIREQ-304.
+* Patron Blocks: Is there an automated patron block on requesting? Refs UIREQ-471.
 
 ## [2.1.0] (IN PROGRESS)
 * Paneheader Actions updates. Refs UIREQ-415.